Refactor: rename ui folder to lcars
